Simple working version that is performant enough to draw in realtime.
This gives us a grid, sets the cursor to a cross and lets us draw.  On
my M1 Max Macbook Pro it seems to be using around 30% CPU to click a
starting point and then wildly sweep the mouse around the canvas
dragging the line.  Also seems to be perfectly usable on my ancient 2011
11" Macbook Air, although I didn't look at CPU utilization.