tools/text_tool.py: Use a TextEntryInspector to get the text to add.
Note that the tkinter.Text widget always includes a '\n' at the end of
the text.  However, the Canvas text object that we draw a string into
doesn't want a trailing '\n' otherwise it starts a new line.  So, we
play silly games here to get the things to match up.